src/controller/requirements.txt
Django==3.0.8
djangorestframework==3.11.0
psycopg2==2.8.6
flake8>=3.6.0,<3.7.0
django-import-export==2.3.0
django-user-messages==0.7.0
djangorestframework-jwt==1.11.0
tablib==0.14.0
pywinrm==0.4.1
paramiko>=2.4
pypsexec==0.1.0
asgiref==3.2.10
defusedxml==0.6.0
diff-match-patch==20200713
django-cors-headers==3.4.0
gunicorn==20.0.4
sqlparse==0.3.1
MarkupPy==1.14
pytz==2020.1
PyYAML==5.3.1
PyJWT==1.7.1
xlwt==1.3.0
openpyxl==3.0.4
odfpy==1.4.1
xlrd==1.2.0
et-xmlfile==1.0.1
jdcal==1.4.1


